+++ azure_monitor/CHANGELOG.md
@@ -1,3 +1,18 @@
+# 2.0.0
+## Added
+- Added support for the content and device alert type.
+- Added support for clientstatus and endpoint events.
+- Added support for JA3 fields in Application events.
+- Added support for invoking mapping validation separately.
+- Added support for log source identifier.
+- Added support for storing access token in the storage.
+## Changed
+- Enhanced the efficiency of database interactions.
+- Updated the mappings for Network events and Audit events.
+## Fixed
+- Fixed CEF transformation for nested JSON fields in data.
+- Fixed ingestion for selected field in JSON format.
+
 # 1.1.0
 ## Added
 - Added Support for the incident event type. To pull and ingest this event type update your CE version to 4.1.0.